refactor(DataMapper): Extract field context menu types and builders into model/service layer
Fixes: https://github.com/KaotoIO/kaoto/issues/3538

- Extract field context menu logic from hooks/components into WrapperActionService methods
- Introduce field-action.ts model collection as a contract
